requirement.txt
opencv-contrib-python==4.4.0.46
moviepy==1.0.3
ffmpeg-python==0.2.0


#cv2 (opencv-python) [4.4.0]
#pyzbar (https://github.com/NaturalHistoryMuseum/pyzbar) [can be installed via pip]
#os, argparse